time stamp

+++ java-spring/src/main/java/com/ditto/example/spring/quickstart/service/Post.java
@@ -6,7 +6,7 @@ public record Post(
         String id,
         String parent,
         String author_id,
-        int time,
+        long time,
         String text,
         String attachment,
         int likes,
